add support for constraints of RULE, LHS, target side of RHS; still need to change the inferface and do debug
git-svn-id: https://joshua.svn.sf.net/svnroot/joshua/trunk@846 0ae5e6b2-d358-4f09-a895-f82f13dd62a4

+++ src/joshua/decoder/segment_file/ConstraintRule.java
@@ -89,5 +89,5 @@ public interface ConstraintRule {
 	 * grammar. We cannot enforce this requirement, but the
 	 * {@link Chart} must throw an error if there is a mismatch.
 	 */
-	double[] features();
+	float[] features();
 }
